Add docker container for production environment of django (#5)
This PR provides docker files to setup gunicorn (for django), nginx (to serve static files and redirect to gunicorn otherwise) and Postgres as database. These are intended to be used later in production.

Note: For production there are still some settings to change in django but the container structure should run now.

* No automigration on docker-compose

This is to avoid accidental override of the database in production. To apply migrations, just use the standard django workflow on the docker database volume.